--- front/src/components/navbar/Navbar.module.css
@@ -1,23 +1,0 @@
-.navList {
-    list-style-type: none;
-    padding: 0;
-    margin: 0;
-    display: flex; 
-    justify-content: space-between; 
-    width: 100%;
-}
-  
-.navList li {
-    flex: 1;
-    margin-right: 20px; 
-}
-  
-.navList li:last-child {
-    margin-right: 0; 
-}
-
-.navList a{
-    font-size: 20px;
-}
-
-
