+++ front/src/components/navbar/Navbar.module.scss
@@ -1,0 +1,77 @@
+@import "../../scss/variables.scss";
+
+.navList {
+  list-style-type: none;
+  padding: 20px 0px 20px 0px;
+  margin: 0;
+  display: flex;
+  justify-content: space-between;
+  width: 100%;
+  position: relative;
+  align-items: center;
+  background-color: $mainColor;
+  font: $mainFont;
+
+  li {
+    flex: 1;
+    margin-right: 20px;
+
+  }
+
+  a {
+    font-size: 20px;
+  }
+}
+
+.searchBar {
+  display: flex;
+  flex: 1;
+  justify-content: center;
+  align-items: center;
+  align-self: center;
+  input[type="text"] {
+    padding: 6px 10px;
+    border-radius: 8px;
+    border: none;
+    margin: 0;
+    width: 80%;
+    max-width: 80%;
+    font-size: 14px;
+  }
+}
+
+
+
+.toRight{
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  position: relative;
+  
+  .dropdown{
+  position: absolute;
+  top: 100%; 
+  right: 0;
+  background-color: white; 
+  border: 1px solid #ccc;
+  padding: 10px;
+  z-index: 1000;
+  margin: 0;
+
+  ul{
+    padding-left: 10px;
+  }
+  li{
+    list-style-type: none;
+    font-size: 14px;
+    
+    a{
+      text-decoration: none;
+      text-align: center;
+      font-size: 14px;
+    }
+  }
+}
+}
+/*/movies/search?*/
+
